Ajustado abordagens de performance do fluxo de item fornecedor

+++ src/main/java/com/br/dreamday/dao/DaoItemFornecedor.java
@@ -10,7 +10,7 @@ public interface DaoItemFornecedor {
     void inserir(ItemFornecedor itemFornecedor);
     void alterar(ItemFornecedor  itemFornecedor);
     void excluirPor(Long idFornecedor, Long idProduto);
-    List<ItemFornecedor> listarPor(String nomeProduto, String nomeFornecedor, BigDecimal valorInicial, BigDecimal valorFinal);
+    List<ItemFornecedor> listarPor(String nomeProduto, String nomeFornecedor, Double valorInicial, Double valorFinal);
     List<ItemFornecedor> listarPor(Long idFornecedor);
     boolean validarRemocaoCategoria(Long idCategoria);
     boolean validarRemocaoFornecedor(Long idFornecedor);
